moved app.js to src
