+++ lambda.js
@@ -1,5 +1,5 @@
 const awsServerlessExpress = require('aws-serverless-express');
-const app = require('./app');
+const app = require('./src/app');
 const server = awsServerlessExpress.createServer(app);
 
 exports.handler = (event, context) => {
