+++ test/api-test.js
@@ -1,6 +1,6 @@
 const request = require('supertest');
 const { expect } = require('chai');
-const app = require('./../app');
+const app = require('./../src/app');
 describe('api', () => {
   it('GET /', (done) => {
     request(app)
